+++ src/commands/zsh/__complete__.ts
@@ -9,6 +9,7 @@ export const zshCompleteCommand = new Command({
   example: 'denvig zsh __complete__ -- denvig services',
   args: [],
   flags: [],
+  acceptsExtraArgs: true,
   handler: async ({ project, extraArgs = [] }) => {
     // Get the words passed after --
     const words = extraArgs